abstract model handers out into a new file
